viewer refactor
