+++ src/main.ts
@@ -1,6 +1,7 @@
 export { AtlasDataset as AtlasProject } from './project.js';
 export { AtlasDataset } from './project.js';
-export { AtlasUser, APIError } from './user.js';
+export { AtlasUser } from './user.js';
+export { APIError, AtlasViewer } from './viewer.js';
 export { AtlasOrganization } from './organization.js';
 export { AtlasProjection } from './projection.js';
 export { AtlasIndex } from './index.js';
