+++ src/project.ts
@@ -1,10 +1,9 @@
 import type { Schema, Table } from 'apache-arrow';
-import type { ApiCallOptions } from './user.js';
 import { tableToIPC, tableFromJSON, tableFromIPC } from 'apache-arrow';
-import { AtlasUser, get_env_user, BaseAtlasClass } from './user.js';
+import { AtlasUser, BaseAtlasClass } from './user.js';
 import { AtlasIndex } from './index.js';
+import { AtlasViewer } from 'viewer.js';
 // get the API key from the node environment
-import { OrganizationProjectInfo } from 'organization.js';
 type UUID = string;
 
 export function load_project(options: Atlas.LoadProjectOptions): AtlasDataset {
@@ -74,7 +73,7 @@ export class AtlasDataset extends BaseAtlasClass<Atlas.ProjectInfo> {
    *
    * @returns An AtlasDataset object.
    */
-  constructor(id: UUID | string, user?: AtlasUser) {
+  constructor(id: UUID | string, user?: AtlasUser | AtlasViewer) {
     super(user);
     // check if id is a valid UUID
 
@@ -116,10 +115,16 @@ export class AtlasDataset extends BaseAtlasClass<Atlas.ProjectInfo> {
     method: 'GET' | 'POST',
     payload: Atlas.Payload = null,
     headers: null | Record<string, string> = null,
-    options: ApiCallOptions = {}
+    options: Atlas.ApiCallOptions = {}
   ) {
     const fixedEndpoint = this._fixEndpointURL(endpoint);
-    return this.user.apiCall(fixedEndpoint, method, payload, headers, options);
+    return this.viewer.apiCall(
+      fixedEndpoint,
+      method,
+      payload,
+      headers,
+      options
+    );
   }
 
   async delete() {
@@ -139,7 +144,7 @@ export class AtlasDataset extends BaseAtlasClass<Atlas.ProjectInfo> {
     return new Promise((resolve, reject) => {
       const interval = setInterval(async () => {
         // Create a new project to clear the cache.
-        const renewed = new AtlasDataset(this.id, this.user);
+        const renewed = new AtlasDataset(this.id, this.viewer);
         const info = (await renewed.info()) as Atlas.ProjectInfo;
         if (info.insert_update_delete_lock === false) {
           clearInterval(interval);
@@ -151,7 +156,7 @@ export class AtlasDataset extends BaseAtlasClass<Atlas.ProjectInfo> {
     });
   }
 
-  endpoint() {
+  protected endpoint() {
     return `/v1/project/${this.id}`;
   }
 
@@ -176,7 +181,7 @@ export class AtlasDataset extends BaseAtlasClass<Atlas.ProjectInfo> {
     }
     const options = { project: this };
     this._indices = atlas_indices.map(
-      (d) => new AtlasIndex(d['id'], this.user, options)
+      (d) => new AtlasIndex(d['id'], this.viewer, options)
     );
     return this._indices;
   }
@@ -268,13 +273,13 @@ export class AtlasDataset extends BaseAtlasClass<Atlas.ProjectInfo> {
       fields
     );
     const id = response as string;
-    return new AtlasIndex(id, this.user, { project: this });
+    return new AtlasIndex(id, this.viewer, { project: this });
   }
 
   async delete_data(ids: string[]): Promise<void> {
     // TODO: untested
     // const info = await this.info
-    await this.user.apiCall('/v1/project/data/delete', 'POST', {
+    await this.viewer.apiCall('/v1/project/data/delete', 'POST', {
       project_id: this.id,
       datum_ids: ids,
     });
