+++ src/projection.ts
@@ -3,6 +3,7 @@ import { BaseAtlasClass } from './user.js';
 import type { AtlasUser } from './user.js';
 import { AtlasDataset } from './project.js';
 import type { AtlasIndex } from './index.js';
+import { AtlasViewer } from 'viewer.js';
 
 export type ProjectGetInfo = Record<string, any>;
 
@@ -92,11 +93,11 @@ export class AtlasProjection extends BaseAtlasClass<ProjectGetInfo> {
 
   constructor(
     public id: UUID,
-    user?: AtlasUser,
+    user?: AtlasUser | AtlasViewer,
     options: ProjectionInitializationOptions = {}
   ) {
     const { project, project_id } = options;
-    super(user || project?.user);
+    super(user || project?.viewer);
 
     if (project_id === undefined && project === undefined) {
       throw new Error('project_id or project is required');
@@ -266,7 +267,7 @@ export class AtlasProjection extends BaseAtlasClass<ProjectGetInfo> {
 
   async project(): Promise<AtlasDataset> {
     if (this._project === undefined) {
-      this._project = new AtlasDataset(this.project_id, this.user);
+      this._project = new AtlasDataset(this.project_id, this.viewer);
     }
     return this._project;
   }
@@ -297,13 +298,13 @@ export class AtlasProjection extends BaseAtlasClass<ProjectGetInfo> {
    * 'public' may be be added in fetching.
    */
   get quadtree_root(): string {
-    const protocol = this.user.apiLocation.startsWith('localhost')
+    const protocol = this.viewer.apiLocation.startsWith('localhost')
       ? 'http'
       : 'https';
-    return `${protocol}://${this.user.apiLocation}/v1/project/${this.project_id}/index/projection/${this.id}/quadtree`;
+    return `${protocol}://${this.viewer.apiLocation}/v1/project/${this.project_id}/index/projection/${this.id}/quadtree`;
   }
 
-  endpoint() {
+  protected endpoint() {
     return `/v1/project/${this.project_id}/index/projection/${this.id}`;
   }
 }
