+++ tsconfig.json
@@ -2,18 +2,12 @@
   "compilerOptions": {
     "moduleResolution": "node",
     "module": "es2020",
-    "lib": [
-      "es2020",
-      "DOM"
-    ],
+    "lib": ["es2020", "DOM"],
     "paths": {
-      "*": [
-        "src/*"
-      ],
-      "@types/*": [
-        "types/*"
-      ]
+      "*": ["src/*"],
+      "@types/*": ["types/*"]
     },
+    "resolveJsonModule": true,
     "target": "es2020",
     "sourceMap": false,
     "esModuleInterop": true,
@@ -26,9 +20,5 @@
     "strict": true,
     "declaration": true
   },
-  "include": [
-    "global.d.ts",
-    "src/global.d.ts",
-    "src/**/*.ts"
-  ]
+  "include": ["global.d.ts", "src/global.d.ts", "src/**/*.ts"]
 }
